pyproject.toml
[tool.poetry]
name = "bluenaas-proxy-local"
version = "0.1.0"
description = ""
authors = ["Blue Brain Project, EPFL"]
readme = "README.md"
packages = [{include = "bluenaas_proxy_local"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.110.0"}
uvicorn = "^0.27.1"
httpx = "^0.27.0"
pyjwt = "^2.8.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.2.2"
mypy = "^1.8.0"
pre-commit = "^3.6.2"
pytest = "^8.0.2"
requests = "^2.31.0"
types-requests = "^2.31.0.20240311"
pytest-asyncio = "0.21.1"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
